CTAMSCommon/src/test/java/org/wuspba/model/BandMemberTest.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */
package org.wuspba.model;

import javax.xml.bind.JAXBContext;
import javax.xml.bind.JAXBException;
import javax.xml.bind.Unmarshaller;
import org.apache.log4j.Logger;
import static org.junit.Assert.*;
import org.junit.Test;

/**
 *
 * @author atrimble
 */
public class BandMemberTest extends AbstractTest {

    private static final Logger LOG = Logger.getLogger(BandMemberTest.class);

    @Test
    public void testBandMember() {
            
        String packageName = CTAMSDocument.class.getPackage().getName();

        try {
            JAXBContext jc = JAXBContext.newInstance(packageName);
            Unmarshaller u = jc.createUnmarshaller();
            Object obj = u.unmarshal(getClass().getResourceAsStream("/bandMember.xml"));

            assertTrue(obj instanceof CTAMSDocument);

            CTAMSDocument ctams = (CTAMSDocument)obj;
            
            assertEquals(ctams.getBandMembers().size(), 1);
            BandMember member = ctams.getBandMembers().get(0);

            assertEquals(member.getId(), andyMember.getId());
            assertEquals(member.getPerson(), andyMember.getPerson());
            assertEquals(member.getType(), andyMember.getType());
            
        } catch (JAXBException ex) {
            LOG.error("Cannot marshal", ex);
            fail();
        }
    }
}
